refactor: trim scheduler time zone migration
