--- .tool-versions
@@ -1,1 +1,0 @@
-elixir 1.18.4-otp-26
