+++ test/jido/agent_server/cron_integration_test.exs
@@ -11,21 +11,7 @@ defmodule JidoTest.AgentServer.CronIntegrationTest do
   alias Jido.Persist
   alias Jido.Signal
   alias Jido.Storage.ETS
-
-  defmodule FailingTimeZoneDatabase do
-    @moduledoc false
-    @behaviour Calendar.TimeZoneDatabase
-
-    @impl true
-    def time_zone_period_from_utc_iso_days(_iso_days, _time_zone) do
-      {:error, :time_zone_not_found}
-    end
-
-    @impl true
-    def time_zone_periods_from_wall_datetime(_datetime, _time_zone) do
-      {:error, :time_zone_not_found}
-    end
-  end
+  alias JidoTest.Support.FailingTimeZoneDatabase
 
   defmodule CronCountAction do
     @moduledoc false
