+++ test/jido/integration/scheduler_integration_test.exs
@@ -14,26 +14,12 @@ defmodule JidoTest.Integration.SchedulerIntegrationTest do
 
   alias Jido.AgentServer
   alias Jido.Signal
+  alias JidoTest.Support.FailingTimeZoneDatabase
 
   @moduletag :integration
   @moduletag :scheduler_integration
   @moduletag capture_log: true
   @moduletag timeout: 20_000
-
-  defmodule FailingTimeZoneDatabase do
-    @moduledoc false
-    @behaviour Calendar.TimeZoneDatabase
-
-    @impl true
-    def time_zone_period_from_utc_iso_days(_iso_days, _time_zone) do
-      {:error, :time_zone_not_found}
-    end
-
-    @impl true
-    def time_zone_periods_from_wall_datetime(_datetime, _time_zone) do
-      {:error, :time_zone_not_found}
-    end
-  end
 
   defmodule PluginTickAction do
     @moduledoc false
